Style : move print_node function to parse
