+++ parse/parse_test.h
@@ -1,19 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   parse_error.c                                      :+:      :+:    :+:   */
+/*   parse_test.h                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: soljeong <soljeong@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/03/26 12:37:19 by soljeong          #+#    #+#             */
-/*   Updated: 2024/03/28 17:11:28 by soljeong         ###   ########.fr       */
+/*   Created: 2024/03/29 18:11:27 by soljeong          #+#    #+#             */
+/*   Updated: 2024/03/29 18:16:42 by soljeong         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#ifndef PARSE_TEST_H
+# define PARSE_TEST_H
 #include "parse.h"
-#include "execute/execute.h"
 
-void	parse_error(void)
-{
-	error_handler(NULL, NULL, "syntax error near unexpected token\n");
-}
+void	print_node(t_token *node);
+#endif PARSE_TEST_H
